Separated out the cookies based user pref impl from gadgets.js. The cookie impl is not production safe and should not be used as a default implementation. However, the samples need something to work and so pull in this file separately.
I replaced the default impl with an empty implementation. At some point in the future we need to replace this with something better.

In the future we should also consider separating all of the gadgets classes into their own files but this needs to go hand in hand with a general improvement to the overall framework. 


gadgets.js by itself is now production safe.




git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/shindig/trunk@640803 13f79535-47bb-0310-9956-ffa450edef68