SHINDIG-202
Subclasses of person will now have all of their fields as well as all of the Person.java fields outputted into json. To fix this I switched the toJson method from using reflection on fields to using methods. 

Eventually we should use a full json to pojo library if we still need this once we restful is in place, but this fixes the bug for now. 




git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/shindig/trunk@649609 13f79535-47bb-0310-9956-ffa450edef68