Adds basic crypto security token logic. The crypto part is courtesy of Bruno Rovagnati.
The crypto should be compatible with the java one (haven't actually
tested yet but it's the same logic).

It introduces 4 new config flags:
st_max_age - max age of the token to be accepted
token_cipher_key & token_hmac_key - keys to use
allow_plaintext_token - allow plain text security token
aka in (john.doe:jane.doe:etc:format)

This last setting should _never_ be true in a real situation but
is required to make the basic code work with the javascript
examples and sample container (which doesn't have crypto tokens).
This might be removed later when i have a good solution for this.



git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/shindig/trunk@652719 13f79535-47bb-0310-9956-ffa450edef68