SHINDIG-203 All of the opensocial model objects have been extracted into interfaces. This gives implementors of the social data apis mroe flexibility.
This code is not fully cleaned up yet because the BeanJsonConverter needs concrete classes in order to translate from json to pojos. Because of this some of the dataservice classes reference ActivityImpl directly and the Activity interface itself has a set MediaItemImpl method on it. There are todos in the code to fix this as soon as we can. 




git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/shindig/trunk@669613 13f79535-47bb-0310-9956-ffa450edef68