Next step in content fetching pipeline improvements for SHINDIG-523 / SHINDIG-517.
This patch primarily refactors HttpResponse to match HttpRequest more closely, with the primary difference being that HttpResponse is immutable (constructed from an HttpResponseBuilder).

Additionally, this patch reverses usage of Multimap for headers and Uri parameters in favor of a Map<String, List<String>>. This is done because the Multimap interface can lead to easy to miss bugs, such as using Sets for the container type. In the future we should probably write a custom container for these objects that meets the needs of Shindig better, such as UriParameter and HttpHeader classes.



git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/shindig/trunk@687932 13f79535-47bb-0310-9956-ffa450edef68