SHINDIG-549
Patch supplied by Adam Winer

replaces error processing that was using in the SPI to using a SocialSpiExcepion (as a runtime exception)

Additionally moved the ResponseItem to opensocial/service as there are no bindings outside
that package and it makes sense for it to be part of the service implementation.

Thanks.


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/shindig/trunk@690036 13f79535-47bb-0310-9956-ffa450edef68